fix possible _non_existent call DoS vulnerability (reported by Christoph Becker)
This patch fixes possible _non_existent call DoS vulnerability. The
problem was observed when sending _non_existent message together with
corrupted target object ID. Actual problem was to try to answer
incorrectly initialized invocation which resulted in assert. The fix
is to initialize invocation properly even in case of _non_existent call
where we're not able to find appropriate OA and which is answered by
the ORB itself. The issue was reported by Christoph Becker.