fix memory leak in MICO codesets initialization
This patch fixes memory leak in MICO codesets initialization. The
problem was caused by the fact that Codeset::_init operation is called
from ORB_init and it did not check if it was already called or not.

darcs-hash:e9541749a72b9e495d9d1ea6d38363f6aa8aa1d3