fix race-condition in GIOP connection while demarshaling object reference (Arne Pajunen)
This patch fixes race-condition in GIOPConn class which happens while
demarshaling returned object reference if the reference points to
different server/connection. In this case demarshaler's code might
call remote _is_a operation (depending on the ORB's isa cache) and
since this is done in reader thread, the code ends in dispatcher
waiting for incoming requests. The problem is that this call is sent
to different connection and also this connection reads incoming
requests. So at the end we do have two GIOPConn objects reading from
the same TCP connection which results in various asserts and
crashes. The solution is to not use static methods for decision if the
thread is the reader, but makes those methods non-static and call them
on the request's own connection. The issue was reported and the patch
is provided by Arne Pajunen