fix static data initialization ordering issue for PI initializers
This patch fixes static data initialization ordering issue which shows
itself in PI implementation in a way that application's initializer is
correctly registered by the ORB and then silently "forgotten" during
the ORB_init call. The problem is that the code made incorrect
assumption about ordering of static data initializations, but such
ordering is not specified by the C++ language and is implementation
dependant. The fix uses construction of the data on its first use
using local static member data trick.