fix to not include os-misc.h in CORBA.h due to compilation issue in application code on the Win32/VC++ platform
This patch changes CondVar::timedwait method implementation to non
inlined one. The reason is a need for removing include of os-misc.h
from the CORBA.h which causes compilation issues on the Win32/VC++
platform in the Qedo application code, i.e. some of the header files
included from os-misc.h declares DOMDocument class while this class is
also declared by the Xerces/C++ which is used in Qedo.