use libtcmalloc_minimal if it is provided by the target OS (scalability improvement)
This patch enhances MICO's configure to detect and use
libtcmalloc_minimal if it is provided by the target OS. It is
scalability improvement as standard OS provided malloc/free library
might not scale well with the number of threads. The feature might be
disabled by using --disable-tcmalloc configure command-line option.
This patch also regenerates config.h.in.