fix destruction of IR Container object to avoid RW lock double-locking issues
This patch fixes destruction of IR Container object in a way it avoids
RW lock double locking issues. The problem here actually was that IR
was using RW lock in recursive manner that means it locked RW lock
several times and then unlock it in reverse order. The problem is that
RW lock as per POSIX spec. does not support recursive locking and in
fact such locking causes undefined behavior. On Linux (Ubuntu 16.04)
this means that second unlock operation crashes with illegal
instruction issue. The issue is caused by invocation of xend isn in
_xend() function which is completely misleading as this isn is defined
in Haswell TSX (transaction memory ISA) extension.