fix to avoid race-condition on resolve_ip call (reported by Wolfgang Schneider)
This patch tries to avoid race-condition on a resolve_ip call on
ORB's IOR template IIOP profile InetAddress object. If we
resolve_ip/resolve_host sooner than when the actual call to server
happen then we avoid the race-condition and do not need to use lock to
protect InetAddress object structure.