fix compilation failure in time service when sys/timeb.h is not available
The patch enhances configure.in to test for sys/timeb.h and if it is
available then defining appropriate CPP define. This is then used in
time service to #ifdef include of sys/timeb.h. Looks like even without
this header the time service is compilable. The problematic platform
here was OpenBSD 6.3-current.