fix IIOPProxy make_conn search for cached connection when SL3 is not in use
This patch fixes IIOPProxy make_conn method search for cached GIOP
connection when SL3 is enabled in build, but not used during the
runtime. In this case both tcpip and tls creds_id variables are 0/NULL
and from cached connection obtained creds_id is empty string. When
this happen we can reuse the connection well.