fix IDL compiler code name collision with GNU C++ >= 7.x (string_literals)
This patch fixes "string_literals" name collision which affects
compiling of code of the IDL compiler with GNU C++ >= 7.x. The problem
is that modern GNU C++ defines string_literals namespace inside its
libstdc++ library. The patch resolves this conflict but as this
involve bison generated files, it also requires regeneration of bison
generated files which will be committed in next patch.