fix assert in POAMediator notify of bind reply
This patch fixes assert in POAMediator notify of bind reply. The
assert was on i != requests.end() which means the requirement was that
the pool of requests sent should not be empty. The problem was that
sometimes where there were several kind of servers involved it may be
empty since we may receive bind reply with LocateHere stat and then we
will be canceling and removing all other outstanding requests but then
the reply from yet another server with LocateUnknown (as the server is
of different type) may come and the assert will happen. Now when
requests queue is empty and bind reply is LocateUnknown we simply
break from the switch statement and return from the method since there
is nothing to do in this case.