fix hangup in POA::activate_object_with_id method
The issue happen when there is ongoing pressure of incoming invocation
performed on the servant and servant is deactivated
and immediately activated again. In this case we wait on condition
variable and we need to check for signal variable value
to protect code from spurious wake ups on the cond var.
The check for signal value was wrong here.

test/mt/stress is also adjusted to test for this issue.

darcs-hash:d47bf7812a7fe591fbc0aba5e3f690eb7bbfb086